Back to long-double
This reverts commit cfb81e6a5bf761b6603a6425e689d82d5f3b046c.

+++ liblwgeom/ptarray.c
@@ -1083,7 +1083,7 @@ ptarray_signed_area(const POINTARRAY *pa)
 	const POINT2D *P1;
 	const POINT2D *P2;
 	const POINT2D *P3;
-	double sum = 0.0;
+	long double sum = 0.0;
 	double x0, x, y1, y2;
 	uint32_t i;
